RL importance-ratio consistency + global_std option + PickScore reward
From debugging why RL on dense rewards (aesthetic/PickScore) wouldn't climb:

- Store the float32 sampled latent as the replay target (was bf16-rounded);
  with the SDE std -> 0 at low sigma, the rounding shifted the target and
  biased the per-step log-prob.
- Match replay's autocast to sampling (cache_enabled=False) in the RL trainers.
- Add advantage_global_std option (FlowGRPO-style) for low-variance rewards.
- Add PickScore reward (yuvalkirstain/PickScore_v1).

KNOWN, still open: even after these, the sampling/replay importance ratio is
~0.92 at epoch 0 in real training (should be 1.0). A controlled diagnostic
shows the ratio is exactly 1.0 when the replay forward matches sampling
(eval mode, no_grad, matched autocast), so the residual bias comes from the
train-mode + grad + gradient-checkpointing replay path. This biases dense-
reward RL; OCR/DRaFT/DiffusionDPO/SFT are unaffected (sparse reward / no
importance ratio).

+++ src/diffusion_tinker/models/sd3_patch.py
@@ -144,10 +144,16 @@ def sd3_sample_with_logprob(
             generator=generator,
         )
 
-        latents = latents.to(dtype)
-
+        # Store the float32 sampled latent as the replay target. log_prob was
+        # computed on this exact value, so replay must see it too - rounding to
+        # bf16 here would shift the target, and since the SDE std shrinks toward
+        # 0 at low sigma, that tiny shift blows up the per-step log-prob and
+        # biases the importance ratio away from 1.0. Round only for the next
+        # step's transformer input.
         all_next_latents.append(latents.detach().cpu())
         all_log_probs.append(log_prob.detach().cpu())
+
+        latents = latents.to(dtype)
 
     decoded = _decode_latents(pipeline, latents)
     images = _tensor_to_pil(decoded)
